Revert "220317 add having step (#20)"
This reverts commit 5c46b7c376faf392961703c0a20456884d6758a9.
